+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.1.1
 ### Added:
  - DateTimeTrait::$nullAllowed : Allows the developer to specify null values or other objects and get null values
- - By default, $nullAllowed is true, to reduce checks.  
+ - By default, $nullAllowed is true, to reduce checks.
+ - Added functions to get an Excel Column (int/string) from a Property  
 
 ## 0.1
  - Initial Version